backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.16.5
aiosqlite==0.22.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
greenlet==3.2.4
anthropic==0.40.0
python-multipart==0.0.18
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
structlog==24.4.0
